@@ -4,6 +4,7 @@
   "description": "Read .2bit files in node or webpack",
   "license": "MIT",
   "main": "dist/index.js",
+  "module": "esm/index.js",
   "repository": "GMOD/twobit-js",
   "author": {
     "name": "Robert Buels",
@@ -14,22 +15,22 @@
     "node": ">=6"
   },
   "files": [
-    "dist"
+    "dist",
+    "esm"
   ],
   "scripts": {
     "test": "jest",
     "coverage": "npm test -- --coverage",
     "lint": "eslint src test",
     "docs": "documentation readme src/twoBitFile.js --section=TwoBitFile",
-    "clean": "rimraf dist",
+    "clean": "rimraf dist esm",
     "prebuild": "npm run docs && npm run clean",
-    "build": "babel src -d dist",
+    "build:esm": "tsc --target es2018 --outDir esm",
+    "build:es5": "tsc --target es5 --outDir dist",
+    "build": "npm run build:esm && npm run build:es5",
     "version": "standard-changelog && git add CHANGELOG.md",
     "prepublishOnly": "npm run lint && npm test && npm run build",
     "postpublish": "git push origin master --follow-tags"
-  },
-  "jest": {
-    "testEnvironment": "node"
   },
   "keywords": [
     "bionode",
@@ -37,28 +38,25 @@
   ],
   "dependencies": {
     "@gmod/binary-parser": "^1.3.5",
-    "fs-extra": "^7.0.0",
+    "es6-promisify": "^7.0.0",
+    "generic-filehandle": "^2.2.1",
     "long": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.3",
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-transform-runtime": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/runtime-corejs2": "^7.4.3",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.7.0",
+    "@types/jest": "^27.0.3",
     "documentation": "^9.3.1",
-    "eslint": "^5.12.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-import": "^2.10.0",
-    "eslint-plugin-prettier": "^3.0.1",
-    "jest": "^24.7.1",
-    "jest-cli": "^24.7.0",
-    "prettier": "^1.17.0",
+    "eslint": "^7.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jest": "^27.4.5",
+    "prettier": "^2.5.1",
     "rimraf": "^2.6.2",
-    "standard-changelog": "^1.0.0"
+    "standard-changelog": "^1.0.0",
+    "ts-jest": "^27.1.1",
+    "typescript": "^4.5.4"
   },
   "publishConfig": {
     "access": "public"
