@@ -1,10 +1,12 @@
-const fs = require('fs')
-const { TwoBitFile } = require('../src')
+//@ts-nocheck
+import fs from 'fs'
+import { TwoBitFile } from '../src'
 
 let extended = xit
 try {
-  if (fs.existsSync(require.resolve(`./data/extended/wheat.2bit`)))
+  if (fs.existsSync(require.resolve(`./data/extended/wheat.2bit`))) {
     extended = it
+  }
 } catch (e) {
   // ignore
   console.log(
